fake date-time service improvements PHR-3776
Changing FakeDateTimeService to fall back to a default implementation returning system time, for simplifying manual testing on the e2e stacks (which is common during code reviews).

Also improving the e2e test infra to be able to set the current time declaratively using the `@CurrentTimestamp` annotation. This is useful because this way the fixed timestamp is _cleared_ before running a test, if it isn't needed.