Add more URIs. GDE-480
Immediate need is to add fhir identifier system
for NL, DE and GB-SCT. For the rest of supported
national IDs only a comment is added.

I couldn't find resolvable URRs neither for NL nor GB-SCT
and even the DE one is only in draft status.
https://www.hl7.org/fhir/datatypes.html#Identifier says
that URLs are preferred, but OIDs are also accepted.