ConfigStorage backed by Spring Environment NEST-1
To make it easier to use the PKB Config classes in Spring Boot applications, I have added an immutable implementation of ConfigStorage that is backed by the Spring Environment rather than our own LayeredLoader. As part of doing this, I have refactored the ConfigStorage implementations somewhat to move more behaviour into the base classes, and introduced interfaces for Mutable/Immutable storages so that we can have multiple implementations.

One consequence of this is that I have changed some core behaviour of MutableConfigStorage to be more consistent with the other implementations. In particular, if you override a configuration key with a null value, it will now behave as the Immutable version would do - i.e. throw a MissingValueException; similarly, if you override a key with a value that won't parse as a particular type, requesting it as that type will result in an exception rather than silently falling back to the original, immutable value. This allowed for significantly cleaner implementation and results in a clearer, more consistent interface. As yet I have found no instances in our current test suite that break as a result of this change; the methods that accept a fallback default value already know how to catch the MissingValueException and return the default, so they are unaffected.