Switch from pubsub based test control to simple(r) http based control.
Pub/sub was a bad fit for the requirements. We synchronously require a response from every participating application before we can proceed with tests. This means a messaging request-reply model is necessary, which apache camel / google pubsub endpoint doesn't support, so we ended up implementing a response aggregator ourselves. This was unnecessarily complex for the task.

Another advantage of this model is that the API for centralised test-control can be the same as the API exposed by each application. Them, the centralised test-control application is reduced to an application registry (which each app calls as it starts up) and a splitting HTTP reverse proxy to forward test-control requests to every participating application.

Add a retrofit based client library for easy consumption.

PHR-8981